Changing #init tests once again
In a previous commit the @game_logic attribute was swapped for a simpler
@key_callback attribute that could be passed more easily to
Window#init. This was predominantly due to finding it difficult to test
all the appropriate method calls were being invoked within
GameEngine#init. This probelem has since been resolved such that
the original method can be used. The key_callback that GameEngine
expects to have available is once again held within a game_logic
variable that will in the future hold other important methods.