Changed format of ShaderProgram.uniforms
Each uniform is now stored as a hash that holds the location and type of
uniform that it is connected to. This should make it easier to
automatically call the correct function later to set the value of a
uniform.

Also tweaked the begin,rescue,ensure blocks so that extra errors
shouldn't be raised while cleaning up.