Implemented the Renderer class
The renderer class currently loads, builds and creates uniforms for
shaders when supplied with a directory.

The shader files need to be named after their desired shaderprogram name
in the renderer followed by the shader type (vertex, geometry or
fragment). For example, to get a shader named ":foo" which consists of a
vertex and fragment shader, the files should be named:
 -- foo_vertex.vs
 -- foo_fragment.fs